WIP - net standard dependencies update to HDF.PInvoke.NETStandard; update code to support HDF 1.10

+++ src/sharpHDF/Structs/Hdf5Identifier.cs
@@ -33,7 +33,7 @@ namespace sharpHDF.Library.Structs
 
         public override int GetHashCode()
         {
-            return Value;
+            return (int)Value;
         }
     }
 }
